LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-opener-policy/no-https-expected.txt

CONSOLE MESSAGE: TypeError: null is not an object (evaluating 'window.opener.functionCalledByOpenee')

Harness Error (TIMEOUT), message = null

TIMEOUT Cross-Origin-Opener-Policy only works over secure contexts Test timed out
PASS Bonus: window.crossOriginIsolated
